scraper/requirements.txt
beautifulsoup4==4.12.3
selenium==4.17.2
requests==2.31.0
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
geopy==2.4.1
psycopg2-binary==2.9.9
schedule==1.2.1
lxml==5.1.0

